rename icon list name
